Refactor settings code to colocate settings with viz
