+++ frontend/src/metabase/visualizations/components/ChartSettings.jsx
@@ -6,7 +6,7 @@ import _ from "underscore";
 import Warnings from "metabase/query_builder/components/Warnings.jsx";
 
 import Visualization from "metabase/visualizations/components/Visualization.jsx"
-import { getSettingsWidgets } from "metabase/lib/visualization_settings";
+import { getSettingsWidgets } from "metabase/visualizations/lib/settings";
 import MetabaseAnalytics from "metabase/lib/analytics";
 import { getVisualizationTransformed } from "metabase/visualizations";
