+++ frontend/src/metabase/visualizations/visualizations/PieChart.jsx
@@ -9,6 +9,7 @@ import ChartWithLegend from "../components/ChartWithLegend.jsx";
 
 import { ChartSettingsError } from "metabase/visualizations/lib/errors";
 import { getFriendlyName } from "metabase/visualizations/lib/utils";
+import { metricSetting, dimensionSetting } from "metabase/visualizations/lib/settings";
 
 import { formatValue } from "metabase/lib/formatting";
 
@@ -47,6 +48,35 @@ export default class PieChart extends Component<*, Props, *> {
         if (!settings["pie.dimension"] || !settings["pie.metric"]) {
             throw new ChartSettingsError("Which columns do want to use?", "Data");
         }
+    }
+
+    static settings = {
+        "pie.dimension": {
+            section: "Data",
+            title: "Dimension",
+            ...dimensionSetting("pie.dimension")
+        },
+        "pie.metric": {
+            section: "Data",
+            title: "Measure",
+            ...metricSetting("pie.metric")
+        },
+        "pie.show_legend": {
+            section: "Display",
+            title: "Show legend",
+            widget: "toggle"
+        },
+        "pie.show_legend_perecent": {
+            section: "Display",
+            title: "Show percentages in legend",
+            widget: "toggle",
+            default: true
+        },
+        "pie.slice_threshold": {
+            section: "Display",
+            title: "Minimum slice percentage",
+            widget: "number"
+        },
     }
 
     componentDidUpdate() {
