+++ frontend/src/metabase/visualizations/visualizations/RowChart.jsx
@@ -5,7 +5,12 @@ import React, { Component, PropTypes } from "react";
 import LineAreaBarChart from "../components/LineAreaBarChart.jsx";
 import { rowRenderer } from "../lib/LineAreaBarRenderer";
 
-export default class BarChart extends LineAreaBarChart {
+import {
+    GRAPH_DATA_SETTINGS,
+    GRAPH_COLORS_SETTINGS
+} from "metabase/visualizations/lib/settings/graph";
+
+export default class RowChart extends LineAreaBarChart {
     static uiName = "Row Chart";
     static identifier = "row";
     static iconName = "horizontal_bar";
@@ -14,4 +19,19 @@ export default class BarChart extends LineAreaBarChart {
     static supportsSeries = false;
 
     static renderer = rowRenderer;
+
+    static settings = {
+        ...GRAPH_DATA_SETTINGS,
+        ...GRAPH_COLORS_SETTINGS
+    }
+}
+
+// rename these settings
+RowChart.settings["graph.metrics"] = {
+    ...RowChart.settings["graph.metrics"],
+    title: "X-axis"
+}
+RowChart.settings["graph.dimensions"] = {
+    ...RowChart.settings["graph.dimensions"],
+    title: "Y-axis"
 }
