+++ frontend/src/metabase/visualizations/visualizations/ScatterPlot.jsx
@@ -5,6 +5,14 @@ import React, { Component, PropTypes } from "react";
 import LineAreaBarChart from "../components/LineAreaBarChart.jsx";
 import { scatterRenderer } from "../lib/LineAreaBarRenderer";
 
+import {
+    GRAPH_DATA_SETTINGS,
+    GRAPH_BUBBLE_SETTINGS,
+    GRAPH_GOAL_SETTINGS,
+    GRAPH_COLORS_SETTINGS,
+    GRAPH_AXIS_SETTINGS
+} from "../lib/settings/graph";
+
 export default class ScatterPlot extends LineAreaBarChart {
     static uiName = "Scatter";
     static identifier = "scatter";
@@ -12,4 +20,12 @@ export default class ScatterPlot extends LineAreaBarChart {
     static noun = "scatter plot";
 
     static renderer = scatterRenderer;
+
+    static settings = {
+        ...GRAPH_DATA_SETTINGS,
+        ...GRAPH_BUBBLE_SETTINGS,
+        ...GRAPH_GOAL_SETTINGS,
+        ...GRAPH_COLORS_SETTINGS,
+        ...GRAPH_AXIS_SETTINGS
+    }
 }
